Move stronghold generation to separate file
