+++ makefile
@@ -30,7 +30,7 @@ release: CFLAGS += -fPIC
 endif
 
 
-libcubiomes: noise.o terrainnoise.o biomes.o layers.o biomenoise.o generator.o finders.o util.o quadbase.o items.o loot_functions.o loot_table_context.o loot_table_parser.o loot_tables.o cjson.o bastion_bridge_1_16_1.o bastion_bridge_1_16_5.o bastion_bridge_1_20.o bastion_other_1_16_1.o bastion_other_1_16_5.o bastion_other_1_20.o bastion_other_1_21_1.o bastion_other_1_21_9.o buried_treasure_1_13.o buried_treasure_1_18.o buried_treasure_1_21_11.o desert_pyramid_1_13.o desert_pyramid_1_20.o desert_pyramid_1_21_6.o desert_pyramid_1_21_9.o desert_pyramid_1_21_11.o end_city_treasure_1_13.o end_city_treasure_1_20.o end_city_treasure_1_21_9.o end_city_treasure_1_21_11.o igloo_chest_1_13.o jungle_temple_1_13.o jungle_temple_1_14.o jungle_temple_1_20.o jungle_temple_1_21_6.o jungle_temple_1_21_9.o jungle_temple_1_21_11.o jungle_temple_dispenser_1_13.o nether_bridge_1_13.o nether_bridge_1_20.o nether_bridge_1_21_9.o pillager_outpost_1_14.o pillager_outpost_1_19_2.o pillager_outpost_1_20.o pillager_outpost_1_21_11.o ruined_portal_1_16_1.o ruined_portal_1_21_5.o shipwreck_map_1_13.o shipwreck_map_1_18.o shipwreck_map_1_20.o shipwreck_map_1_21_11.o shipwreck_supply_1_13.o shipwreck_supply_1_14.o shipwreck_supply_1_17.o shipwreck_supply_1_20.o shipwreck_supply_1_21_11.o shipwreck_treasure_1_13.o shipwreck_treasure_1_20.o shipwreck_treasure_1_21_11.o
+libcubiomes: noise.o terrainnoise.o biomes.o layers.o biomenoise.o generator.o finders.o util.o quadbase.o stronghold.o items.o loot_functions.o loot_table_context.o loot_table_parser.o loot_tables.o cjson.o bastion_bridge_1_16_1.o bastion_bridge_1_16_5.o bastion_bridge_1_20.o bastion_other_1_16_1.o bastion_other_1_16_5.o bastion_other_1_20.o bastion_other_1_21_1.o bastion_other_1_21_9.o buried_treasure_1_13.o buried_treasure_1_18.o buried_treasure_1_21_11.o desert_pyramid_1_13.o desert_pyramid_1_20.o desert_pyramid_1_21_6.o desert_pyramid_1_21_9.o desert_pyramid_1_21_11.o end_city_treasure_1_13.o end_city_treasure_1_20.o end_city_treasure_1_21_9.o end_city_treasure_1_21_11.o igloo_chest_1_13.o jungle_temple_1_13.o jungle_temple_1_14.o jungle_temple_1_20.o jungle_temple_1_21_6.o jungle_temple_1_21_9.o jungle_temple_1_21_11.o jungle_temple_dispenser_1_13.o nether_bridge_1_13.o nether_bridge_1_20.o nether_bridge_1_21_9.o pillager_outpost_1_14.o pillager_outpost_1_19_2.o pillager_outpost_1_20.o pillager_outpost_1_21_11.o ruined_portal_1_16_1.o ruined_portal_1_21_5.o shipwreck_map_1_13.o shipwreck_map_1_18.o shipwreck_map_1_20.o shipwreck_map_1_21_11.o shipwreck_supply_1_13.o shipwreck_supply_1_14.o shipwreck_supply_1_17.o shipwreck_supply_1_20.o shipwreck_supply_1_21_11.o shipwreck_treasure_1_13.o shipwreck_treasure_1_20.o shipwreck_treasure_1_21_11.o
 	$(AR) $(ARFLAGS) libcubiomes.a $^
 
 finders.o: finders.c finders.h
@@ -61,6 +61,9 @@ util.o: util.c util.h
 	$(CC) -c $(CFLAGS) $<
 
 quadbase.o: quadbase.c quadbase.h
+	$(CC) -c $(CFLAGS) $<
+
+stronghold.o: features/stronghold.c features/stronghold.h
 	$(CC) -c $(CFLAGS) $<
 
 items.o: loot/items.c loot/items.h
